Refactoring. Adding process message initial version.
