+++ source/Roy.Logging/Roy.Logging/Domain/Communication/ProcessMessage.cs
@@ -1,0 +1,26 @@
+﻿using Avalon.Base.Extension.Collections;
+
+namespace Roy.Logging.Domain.Communication;
+
+/// <summary>
+/// Message returned by the logging service.
+/// </summary>
+internal class ProcessMessage
+{
+    /// <summary>
+    /// Flag that indicates whether the process succeeded.
+    /// </summary>
+    public bool Succeed => this.Errors.HasElements();
+    /// <summary>
+    /// Errors throw while saving/logging the object.
+    /// </summary>
+    public List<Exception>  Errors { get; set; }
+
+    /// <summary>
+    /// Constructor
+    /// </summary>
+    public ProcessMessage()
+    {
+        this.Errors = new List<Exception>();
+    }
+}
